dom/workers/test/test_location.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Location
-->
<head>
  <title>Test for DOM Worker Location</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" language="javascript">

  var thisFilename = "test_location.html";
  var workerFilename = "location_worker.js";

  var href = window.location.href
  var queryPos = href.lastIndexOf(window.location.search);
  var baseHref = href.substr(0, href.substr(0, queryPos).lastIndexOf("/") + 1);

  var path = window.location.pathname;
  var basePath = path.substr(0, path.lastIndexOf("/") + 1);

  var strings = {
    "toString": baseHref + workerFilename,
    "href": baseHref + workerFilename,
    "protocol": window.location.protocol,
    "host": window.location.host,
    "hostname": window.location.hostname,
    "port": window.location.port,
    "pathname": basePath + workerFilename,
    "search": "",
    "hash": "",
    "origin": "http://mochi.test:8888"
  };

  var lastSlash = href.substr(0, queryPos).lastIndexOf("/") + 1;
  is(thisFilename,
     href.substr(lastSlash, queryPos - lastSlash),
     "Correct filename ");

  var worker = new Worker(workerFilename);

  worker.onmessage = function(event) {
    if (event.data.string == "testfinished") {
      SimpleTest.finish();
      return;
    }
    ok(event.data.string in strings, event.data.string);
    is(event.data.value, strings[event.data.string], event.data.string);
  };

  worker.onerror = function(event) {
    ok(false, "Worker had an error: " + event.message);
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
